leer_sensor/requirements.txt
matplotlib
pyserial 
time 